feat: sketch of a real plugin system
