+++ cloudvolume/mesh.py
@@ -1,0 +1,287 @@
+import struct
+
+import numpy as np
+
+from .exceptions import MeshDecodeError
+from .lib import yellow
+
+NOTICE = {
+  'vertices': 0,
+  'num_vertices': 0,
+  'faces': 0,
+}
+
+def deprecation_notice(key):
+  if NOTICE[key] < 1:
+    print(yellow("""
+  Deprecation Notice: Meshes, formerly dicts, are now PrecomputedMesh objects
+  as of CloudVolume 0.51.0. 
+
+  Please change mesh['{}'] to mesh.{}
+  """.format(key, key)))
+    NOTICE[key] += 1
+
+class Mesh(object):
+  """
+  Represents the vertices, faces, and normals of a mesh
+  as numpy arrays.
+
+  class PrecomputedMesh:
+    ndarray[float32, ndim=2] self.vertices: [ [x,y,z], ... ]
+    ndarray[uint32,  ndim=2] self.faces:    [ [v1,v2,v3], ... ]
+    ndarray[float32, ndim=2] self.normals:  [ [nx,ny,nz], ... ]
+
+  """
+  def __init__(
+    self, vertices, faces, normals=None, 
+    segid=None, encoding_type=None, encoding_options=None
+  ):
+    self.vertices = np.array(vertices, dtype=np.float32)
+    self.faces = np.array(faces, dtype=np.uint32)
+
+    if normals is None:
+      self.normals = np.array([], dtype=np.float32).reshape((0,3))
+    else:
+      self.normals = np.array(normals, dtype=np.float32)
+
+    self.segid = segid
+    self.encoding_type = encoding_type
+    self.encoding_options = encoding_options 
+
+  def __len__(self):
+    return self.vertices.shape[0]
+
+  def __eq__(self, other):
+    """Tests strict equality between two meshes."""
+
+    no_self_normals = self.normals is None or self.normals.size == 0
+    no_other_normals = other.normals is None or other.normals.size == 0
+
+    if no_self_normals != no_other_normals:
+      return False
+       
+    equality = np.all(self.vertices == other.vertices) \
+      and np.all(self.faces == other.faces)
+
+    if no_self_normals:
+      return equality
+
+    return (equality and np.all(self.normals == other.normals))
+
+  def __repr__(self):
+    return "Mesh(vertices<{}>, faces<{}>, normals<{}>)".format(
+      self.vertices.shape[0], self.faces.shape[0], self.normals.shape[0]
+    )
+
+  def __getitem__(self, key):
+    val = None 
+    if key == 'vertices':
+      val = self.vertices
+    elif key == 'num_vertices':
+      val = len(self)
+    elif key == 'faces':
+      val = self.faces
+    else:
+      raise KeyError("{} not found.".format(key))
+
+    deprecation_notice(key)
+    return val
+
+  def empty(self):
+    return self.vertices.size == 0 or self.faces.size == 0
+
+  def clone(self):
+    return Mesh(np.copy(self.vertices), np.copy(self.faces), np.copy(self.normals))
+
+  @classmethod
+  def concatenate(cls, *meshes):
+    vertex_ct = np.zeros(len(meshes) + 1, np.uint32)
+    vertex_ct[1:] = np.cumsum([ len(mesh) for mesh in meshes ])
+
+    vertices = np.concatenate([ mesh.vertices for mesh in meshes ])
+    
+    faces = np.concatenate([ 
+      mesh.faces + vertex_ct[i] for i, mesh in enumerate(meshes) 
+    ])
+
+    normals = np.concatenate([ mesh.normals for mesh in meshes ])
+
+    return Mesh(vertices, faces, normals)
+
+  def consolidate(self):
+    """Remove duplicate vertices and faces. Returns a new mesh object."""
+    if self.empty():
+      return Mesh([], [], normals=None)
+
+    vertices = self.vertices
+    faces = self.faces
+    normals = self.normals
+
+    eff_verts, uniq_idx, idx_representative = np.unique(
+      vertices, axis=0, return_index=True, return_inverse=True
+    )
+
+    face_vector_map = np.vectorize(lambda x: idx_representative[x])
+    eff_faces = face_vector_map(faces)
+    eff_faces = np.unique(eff_faces, axis=0)
+
+    # normal_vector_map = np.vectorize(lambda idx: normals[idx])
+    # eff_normals = normal_vector_map(uniq_idx)
+
+    return Mesh(eff_verts, eff_faces, None, segid=self.segid)
+
+  @classmethod
+  def from_precomputed(self, binary):
+    """
+    Mesh from_precomputed(self, binary)
+
+    Decode a Precomputed format mesh from a byte string.
+    
+    Format:
+      uint32        Nv * float32 * 3   uint32 * 3 until end
+      Nv            (x,y,z)            (v1,v2,v2)
+      N Vertices    Vertices           Faces
+    """
+    num_vertices = struct.unpack("=I", binary[0:4])[0]
+    try:
+      # count=-1 means all data in buffer
+      vertices = np.frombuffer(binary, dtype=np.float32, count=3*num_vertices, offset=4)
+      faces = np.frombuffer(binary, dtype=np.uint32, count=-1, offset=(4 + 12 * num_vertices)) 
+    except ValueError:
+      raise MeshDecodeError("""
+        The input buffer is too small for the Precomputed format.
+        Minimum Bytes: {} 
+        Actual Bytes: {}
+      """.format(4 + 4 * num_vertices, len(binary)))
+
+    vertices = vertices.reshape(num_vertices, 3)
+    faces = faces.reshape(faces.size // 3, 3)
+
+    return Mesh(
+      vertices, faces, normals=None, 
+      encoding_type='precomputed'
+    )
+
+  def to_precomputed(self):
+    """
+    bytes to_precomputed(self)
+
+    Convert mesh into binary format compatible with Neuroglancer.
+    Does not preserve normals.
+    """
+    vertex_index_format = [
+      np.uint32(self.vertices.shape[0]), # Number of vertices (3 coordinates)
+      self.vertices,
+      self.faces
+    ]
+    return b''.join([ array.tobytes('C') for array in vertex_index_format ])
+
+  @classmethod
+  def from_obj(self, text):
+    """Given a string representing a Wavefront OBJ file, decode to a Mesh."""
+
+    vertices = []
+    faces = []
+    normals = []
+
+    if type(text) is bytes:
+      text = text.decode('utf8')
+
+    for line in text.split('\n'):
+      line = line.strip()
+      if len(line) == 0:
+        continue
+      elif line[0] == '#':
+        continue
+      elif line[0] == 'f':
+        if line.find('/') != -1:
+          # e.g. f 6092/2095/6079 6087/2092/6075 6088/2097/6081
+          (v1, vt1, vn1, v2, vt2, vn2, v3, vt3, vn3) = re.match(r'f\s+(\d+)/(\d*)/(\d+)\s+(\d+)/(\d*)/(\d+)\s+(\d+)/(\d*)/(\d+)', line).groups()
+        else:
+          (v1, v2, v3) = re.match(r'f\s+(\d+)\s+(\d+)\s+(\d+)', line).groups()
+        faces.append( (int(v1), int(v2), int(v3)) )
+      elif line[0] == 'v':
+        if line[1] == 't': # vertex textures not supported
+          # e.g. vt 0.351192 0.337058
+          continue 
+        elif line[1] == 'n': # vertex normals
+          # e.g. vn 0.992266 -0.033290 -0.119585
+          (n1, n2, n3) = re.match(r'vn\s+([-\d\.]+)\s+([-\d\.]+)\s+([-\d\.]+)', line).groups()
+          normals.append( (float(n1), float(n2), float(n3)) )
+        else:
+          # e.g. v -0.317868 -0.000526 -0.251834
+          (v1, v2, v3) = re.match(r'v\s+([-\d\.]+)\s+([-\d\.]+)\s+([-\d\.]+)', line).groups()
+          vertices.append( (float(v1), float(v2), float(v3)) )
+
+    vertices = np.array(vertices, dtype=np.float32)
+    faces = np.array(faces, dtype=np.uint32)
+    normals = np.array(normals, dtype=np.float32)
+
+    return Mesh(vertices, faces - 1, normals)
+
+  def to_obj(self):
+    """Return a string representing a .obj file."""
+    objdata = []
+    objdata += [ 'v {:.5f} {:.5f} {:.5f}'.format(*vertex) for vertex in self.vertices ]
+    objdata += [ 'f {} {} {}'.format(*face) for face in (self.faces+1) ] # obj is 1 indexed
+    objdata = '\n'.join(objdata) + '\n'
+    return objdata.encode('utf8')
+
+  def to_ply(self):
+    """
+    Return a bytearray in .ply format, 
+    a more compact format than .obj that's still widely readable.
+    """
+    vertexct = self.vertices.shape[0]
+    trianglect = self.faces.shape[0]
+
+    # Header
+    plydata = bytearray("""ply
+format binary_little_endian 1.0
+element vertex {}
+property float x
+property float y
+property float z
+element face {}
+property list int int vertex_indices
+end_header
+""".format(vertexct, trianglect).encode('utf8'))
+
+    # Vertex data (x y z): "fff" 
+    plydata.extend(self.vertices.tobytes('C'))
+
+    # Faces (3 f1 f2 f3): "3iii" 
+    plydata.extend(
+      np.insert(self.faces, 0, 3, axis=1).tobytes('C')
+    )
+
+    return plydata
+
+  @classmethod
+  def from_draco(cls, binary):
+    import DracoPy
+
+    try:
+      mesh_object = DracoPy.decode_buffer_to_mesh(binary)
+      vertices = np.array(mesh_object.points).astype(np.float32)
+      faces = np.array(mesh_object.faces).astype(np.uint32)
+    except ValueError:
+      raise MeshDecodeError("Not a valid draco mesh.")
+
+    Nv = len(vertices)
+    Nf = len(faces)
+
+    if Nv % 3 != 0: 
+      raise MeshDecodeError("Mesh vertices not 3D. Cannot decode. Num. vertices: {}".format(Nv))
+
+    if Nf % 3 != 0:
+      raise MeshDecodeError("Faces are not sets of triples. Cannot decode. Num. faces: {}".format(Nf))
+
+    vertices = vertices.reshape(Nv // 3, 3)
+    faces = faces.reshape(Nf // 3, 3)
+
+    return Mesh(
+      vertices, faces, normals=None,
+      encoding_type='draco', 
+      encoding_options=mesh_object.encoding_options
+    )
